third_party/WebKit/LayoutTests/platform/win/http/tests/security/referrerPolicyHeader/referrer-policy-header-on-cross-origin-redirect-response.https-expected.txt

CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/referrerPolicyHeader/referrer-policy-header-on-cross-origin-redirect-response.https.html' was loaded over HTTPS, but requested an insecure resource 'http://example.test:8000/security/referrerPolicyHeader/resources/referrer-and-host.php'. This content should also be served over HTTPS.
CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/referrerPolicyHeader/referrer-policy-header-on-cross-origin-redirect-response.https.html' was loaded over HTTPS, but requested an insecure resource 'http://example.test:8000/security/referrerPolicyHeader/resources/referrer-and-host.php'. This content should also be served over HTTPS.
This is a testharness.js-based test.
PASS Untitled 
Harness: the test ran to completion.
